Break util file into fake-redis-helper files.
